FIXED: multiples commas in arguments

+++ asm/include/asm.h
@@ -51,7 +51,7 @@ int check_size(char **params, args_type_t *type, list_t *list);
 char *get_command_name(char *buffer, size_t adv);
 
     // get_command_params.c
-int get_param_len(char *buffer, size_t *adv);
+size_t get_param_len(char *buffer, size_t *adv);
 char *get_one_param(char *buffer, size_t *adv);
 char **get_command_params(char *buffer, size_t adv);
